requirements.txt
streamlit_chat
langchain
PyPDF2
langchain_google_genai
google-generativeai
langchain_community
bs4